Koombea/requirements.txt
appium-python-client==4.5.0
behave==1.2.6
selenium==4.28.1
